fix: 循環参照の問題と、リレーションシップを適切に初期化できない問題を修正
