+++ app/models/_message_registry.py
@@ -1,0 +1,187 @@
+"""リレーションシップ関係で別モジュールで定義すると循環参照などの問題が発生するMessage関連のモデルを定義.
+
+各モデルは別モジュールからエクスポートし、定義部分を分離しているように見せる.
+"""
+
+from datetime import datetime
+
+from pydantic import BaseModel
+from sqlalchemy.orm import relationship
+from sqlmodel import Column, Field, ForeignKey, Integer, Relationship, SQLModel
+
+from models.common import BaseSQLModel
+
+
+class BaseMessage(BaseModel):
+    """MESSAGES のベースモデル.
+
+    vendor_id: 取得元クライアント(Vendor)へのFK
+    folder_id: 所属フォルダ(Folder)へのFK (NULL可 / 移動で更新可)
+    """
+
+    rfc822_message_id: str
+    subject: str
+    date_sent: datetime | None = None
+    date_received: datetime
+    in_reply_to: str | None = None
+    references_list: str | None = None
+    is_read: bool = False
+    is_replied: bool = False
+    is_flagged: bool = False
+    is_forwarded: bool = False
+    vendor_id: int
+    folder_id: int | None = None
+
+
+class Message(BaseMessage, BaseSQLModel, table=True):
+    """MESSAGES テーブルモデル."""
+
+    rfc822_message_id: str = Field(unique=True, index=True)
+    date_received: datetime = Field(index=True)
+
+    vendor_id: int = Field(sa_column=Column(Integer, ForeignKey("vendor.id", ondelete="RESTRICT"), nullable=False))
+    folder_id: int | None = Field(
+        default=None,
+        sa_column=Column(Integer, ForeignKey("folder.id", ondelete="SET NULL"), nullable=True),
+    )
+
+
+class MessageCreate(BaseMessage):
+    """作成用モデル."""
+
+
+class MessageRead(BaseMessage, BaseSQLModel):
+    """読み取り用モデル."""
+
+
+class MessageUpdate(SQLModel):
+    """更新用モデル(更新可のみ)."""
+
+    is_read: bool | None = None
+    is_replied: bool | None = None
+    is_flagged: bool | None = None
+    is_forwarded: bool | None = None
+    folder_id: int | None = None  # フォルダ移動を許可
+
+
+class BaseFolder(BaseModel):
+    """FOLDERS ベースモデル."""
+
+    name: str
+    system_type: str | None = None  # 予約済みフォルダタイプ (Inbox, Trash など)
+
+
+class Folder(BaseFolder, BaseSQLModel, table=True):
+    name: str = Field(unique=True, index=True)
+    system_type: str | None = Field(default=None, unique=True, index=True)
+
+    messages: list[Message] = Relationship(back_populates="folder")
+
+
+class FolderCreate(BaseFolder):
+    pass
+
+
+class FolderRead(BaseFolder, BaseSQLModel):
+    pass
+
+
+class FolderUpdate(SQLModel):
+    name: str | None = None
+
+
+class BaseSummary(BaseModel):
+    """サマリのベースモデル.
+
+    Attributes:
+    message_id (int): メッセージID.
+        content (str): サマリ内容.
+    """
+
+    message_id: int
+    content: str
+
+
+class Summary(BaseSummary, BaseSQLModel, table=True):
+    """サマリモデル."""
+
+    message_id: int = Field(
+        default=None,
+        sa_column=Column(
+            Integer,
+            ForeignKey("message.id", ondelete="CASCADE"),
+            unique=True,
+            index=True,
+        ),
+    )
+
+    message: list[Message] = Relationship(back_populates="summary")
+
+
+class SummaryCreate(BaseSummary):
+    """サマリ作成用モデル.
+
+    アプリ内にサマリを新規登録する際に使用されるモデル。
+
+    Attributes:
+        message_id (str): メッセージID.
+        content (str): サマリ内容.
+    """
+
+
+class SummaryRead(BaseSummary, BaseSQLModel):
+    """サマリ読み取り用モデル.
+
+    アプリ内でサマリを読み取る際に使用されるモデル。
+
+    Attributes:
+        id (int): サマリID.
+        message_id (str): メッセージID.
+        content (str): サマリ内容.
+        created_at (datetime): 作成日時.
+        updated_at (datetime): 更新日時.
+    """
+
+
+class SummaryUpdate(SQLModel):
+    """サマリ更新用モデル.
+
+    アプリ内でサマリを更新する際に使用されるモデル。
+
+    Attributes:
+        content (str): サマリ内容.
+    """
+
+    content: str | None = None
+
+
+class BaseVendor(BaseModel):
+    """VENDORS のベースモデル."""
+
+    name: str
+
+
+class Vendor(BaseVendor, BaseSQLModel, table=True):
+    """取得元クライアント/サービスベンダー."""
+
+    name: str = Field(unique=True, index=True)
+
+    messages: list[Message] = Relationship(back_populates="vendor")
+
+
+class VendorCreate(BaseVendor):
+    pass
+
+
+class VendorRead(BaseVendor, BaseSQLModel):
+    pass
+
+
+class VendorUpdate(BaseModel):
+    # name は更新不可ポリシーなので空 (将来追加する場合用意)
+    pass
+
+
+Message.vendor = relationship("Vendor", back_populates="messages")
+Message.folder = relationship("Folder", back_populates="messages")
+Message.summary = relationship("Summary", back_populates="message")
